pyproject.toml
[tool.poetry]
name = "mashelper-backend"
version = "0.1.0"
description = ""
authors = ["Harish <harishpoctor@gmail.com>"]
readme = "README.md"
packages = [{include = "mashelper_backend"}]

[tool.poetry.dependencies]
python = "^3.7"
Django = "3.2"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
gunicorn = "^20.1.0"
dj-database-url = "0.5"
django-debug-toolbar = "3.4.0"
djangorestframework = "^3.13.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
